+++ components/src/main/java/com/mercadolibre/android/andesui/tooltip/factory/AndesTooltipAttrs.kt
@@ -4,6 +4,7 @@ import com.mercadolibre.android.andesui.tooltip.actions.AndesTooltipAction
 import com.mercadolibre.android.andesui.tooltip.actions.AndesTooltipLinkAction
 import com.mercadolibre.android.andesui.tooltip.style.AndesTooltipStyle
 import com.mercadolibre.android.andesui.tooltip.location.AndesTooltipLocation
+import com.mercadolibre.android.andesui.tooltip.style.AndesTooltipSize
 
 internal data class AndesTooltipAttrs(
     val style: AndesTooltipStyle,
@@ -13,5 +14,6 @@ internal data class AndesTooltipAttrs(
     val mainAction: AndesTooltipAction? = null,
     val secondaryAction: AndesTooltipAction? = null,
     val linkAction: AndesTooltipLinkAction? = null,
-    val tooltipLocation: AndesTooltipLocation
+    val tooltipLocation: AndesTooltipLocation,
+    val andesTooltipSize: AndesTooltipSize
 )