+++ components/src/main/java/com/mercadolibre/android/andesui/tooltip/factory/Constants.kt
@@ -1,0 +1,5 @@
+package com.mercadolibre.android.andesui.tooltip.factory
+
+internal object Constants {
+    const val GENERIC_X_VALUE = 0F
+}