+++ components/src/test/java/com/mercadolibre/android/andesui/tooltip/actions/AndesTooltipCases.kt
@@ -2,6 +2,7 @@ package com.mercadolibre.android.andesui.tooltip.actions
 
 import com.mercadolibre.android.andesui.button.hierarchy.AndesButtonHierarchy
 import com.mercadolibre.android.andesui.tooltip.location.AndesTooltipLocation
+import com.mercadolibre.android.andesui.tooltip.style.AndesTooltipSize
 import com.mercadolibre.android.andesui.tooltip.style.AndesTooltipStyle
 
 val tooltipNoAction: Array<Any?> = arrayOf(
@@ -12,7 +13,8 @@ val tooltipNoAction: Array<Any?> = arrayOf(
         AndesTooltipLocation.LEFT,
         null,
         null,
-        null
+        null,
+        AndesTooltipSize.DYNAMIC
 )
 val tooltipLoud = arrayOf(
         AndesTooltipStyle.LIGHT,
@@ -22,7 +24,8 @@ val tooltipLoud = arrayOf(
         AndesTooltipLocation.TOP,
         AndesTooltipAction("main action", AndesButtonHierarchy.LOUD, { _, _ -> }),
         null,
-        null
+        null,
+        AndesTooltipSize.DYNAMIC
 )
 
 val tooltipLoudAndQuiet = arrayOf(
@@ -33,7 +36,8 @@ val tooltipLoudAndQuiet = arrayOf(
         AndesTooltipLocation.TOP,
         AndesTooltipAction("main action", AndesButtonHierarchy.LOUD, { _, _ -> }),
         AndesTooltipAction("secondary action", AndesButtonHierarchy.QUIET, { _, _ -> }),
-        null
+        null,
+        AndesTooltipSize.DYNAMIC
 )
 
 val tooltipLoudAndTransparent = arrayOf(
@@ -44,7 +48,8 @@ val tooltipLoudAndTransparent = arrayOf(
         AndesTooltipLocation.TOP,
         AndesTooltipAction("main action", AndesButtonHierarchy.LOUD, { _, _ -> }),
         AndesTooltipAction("secondary action", AndesButtonHierarchy.TRANSPARENT, { _, _ -> }),
-        null
+        null,
+        AndesTooltipSize.DYNAMIC
 )
 
 val tooltipLink = arrayOf(
@@ -55,5 +60,18 @@ val tooltipLink = arrayOf(
         AndesTooltipLocation.TOP,
         null,
         null,
-        AndesTooltipLinkAction("link action", { _, _ -> })
+        AndesTooltipLinkAction("link action", { _, _ -> }),
+        AndesTooltipSize.DYNAMIC
+)
+
+val tooltipFullSize = arrayOf(
+    AndesTooltipStyle.LIGHT,
+    "title",
+    "body",
+    true,
+    AndesTooltipLocation.BOTTOM,
+    AndesTooltipAction("main action", AndesButtonHierarchy.LOUD, { _, _ -> }),
+    null,
+    null,
+    AndesTooltipSize.FULL_SIZE
 )